refactor: refactor: decompose lib/plugins/discord.ts (1259 lines) into focused mod (#226)
